Teste com java puro
