+++ css/index.css
@@ -1,0 +1,100 @@
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+}
+
+.App {
+  text-align: center;
+  
+}
+.App-header{
+  background: linear-gradient(180deg, #ff8800,#FFB917);
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+.hide{
+  display: none;
+}
+
+
+h1 {height: 25px;}
+
+.data-Weather {
+  background-color: aliceblue;
+  color: rgb(59, 59, 59);
+  box-shadow: #371801 0px 0px 3px 0px;
+  width: 500px;
+  height: 700px;
+  margin-top: 3%;
+}
+
+.orange-line {
+  background-color: orange;
+  height: 1px; 
+  width: 65%; 
+  margin: 35px auto 0px auto; 
+
+}  
+.dia-Sem{
+  display: inline-block;
+  justify-content: space-around;
+  /*HERE*/
+}
+.Div-search {
+  align-items: center;
+  justify-content: center;
+  height: 35px;
+  width: 400px;
+}
+
+.Input {
+  border: hidden;
+  width: 250px;
+  height: 35px;
+  background-color: white;
+  padding: 0%;
+}
+
+.button {
+  border-left: hidden;
+  background-color: white;
+  height: 35px;
+  width: 35px;
+  border:  0px;
+  cursor: pointer;
+  padding: 0%;
+}
+
+.White-line {
+  background-color: white;
+  height: 1px; 
+  width: 65%; 
+  margin: 35px auto 0px auto; 
+
+}  
+
+h2 {
+  margin-top: 10px;
+}
+
+.Capitais{
+  /*Capitais que aparecem*/
+   /*Capitais que aparecem*/
+}
