--- public/Fonts.txt
@@ -1,5 +1,0 @@
-Fonts used in this design:
-
-Font: Montserrat
-Author: Julieta Ulanovsky,Sol Matas,Juan Pablo del Peral,Jacques Le Bailly
-Url: https://fonts.google.com/specimen/Montserrat
